Replaced some floats with doubles

+++ src/hitable.h
@@ -36,7 +36,7 @@ struct HitRecord
 class Hitable
 {
   public:
-    virtual bool hit(const Ray& r, float t_min, float t_max, HitRecord& rec) const = 0;
+    virtual bool hit(const Ray& r, double t_min, double t_max, HitRecord& rec) const = 0;
 };
 
 #endif
